test: move group-photo symlink cases out of the 600-line file
